fusetest: write to tests cp from outside mount to inside
